fix: make size arithmetic explicit in buffer calculations
Use size_t arithmetic for buffer sizes, offsets, and allocation counts where
values are used as memory sizes or pointer indexes.

This avoids implicit-widening of multiplication results reported by
clang-tidy and makes the intended arithmetic width explicit. The affected
code now performs multiplications in the target size domain instead of first
multiplying in int or unsigned int and widening the result afterwards.

Also align signed and unsigned character comparisons in field-value parsing
to avoid implementation-dependent char comparisons.

These changes reduce static-analysis noise while preserving the existing
runtime behavior.

+++ nsd/adpeval.c
@@ -394,7 +394,7 @@ NsAdpReset(NsInterp *itPtr)
         itPtr->adp.bufsize = itPtr->servPtr->adp.bufsize;
         itPtr->adp.flags = itPtr->servPtr->adp.flags;
     } else {
-        itPtr->adp.bufsize = 1024u * 1000u;
+        itPtr->adp.bufsize = (size_t)1024u * 1000u;
         itPtr->adp.flags = 0u;
     }
     Tcl_DStringSetLength(&itPtr->adp.output, 0);
